requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
beautifulsoup4==4.12.2
blinker==1.9.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cryptography==46.0.1
fastapi==0.117.1
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
frozenlist==1.7.0
greenlet==3.2.4
gunicorn==21.2.0
h11==0.16.0
http_ece==1.2.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.6.4
outcome==1.3.0.post0
packaging==25.0
pillow==10.4.0
propcache==0.3.2
py-vapid==1.9.2
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pywebpush==1.14.0
requests==2.31.0
lxml>=5.2.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
SQLAlchemy==2.0.43
starlette==0.48.0
trio==0.30.0
trio-websocket==0.12.2
ttkbootstrap==1.14.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.36.0
webdriver-manager==4.0.1
websocket-client==1.8.0
Werkzeug==3.1.3
winotify==1.1.0
wsproto==1.2.0
yarl==1.20.1
pytest==7.4.3
psycopg[binary]>=3.1.0
alembic>=1.12
playwright>=1.47.0
pdfminer.six>=20221105
